[*] Refactor folder naming: typings -> models
