+++ scripts/build.sh
@@ -8,5 +8,5 @@ openapi-generator-cli generate \
 		--additional-properties=modelPropertyNaming=original \
 		--additional-properties=helperFunctions=./templates/helpers/camelCase.ts
 
-cp build/model/* src/zalopay/typings
+cp build/model/* src/zalopay/models
 rm -r build
