+++ src/zalopay/models/agreementBalanceResponseDataInner.ts
