+++ src/zalopay/models/agreementBindResponse.ts
