+++ src/zalopay/models/agreementQueryResponse.ts
