+++ src/zalopay/models/agreementQueryResponseData.ts
