+++ src/zalopay/models/agreementQueryUserRequest.ts
