+++ src/zalopay/models/agreementQueryUserResponse.ts
