+++ src/zalopay/models/oAQueryOrderResponse.ts
